+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Update orient (1.7.8 > 1.7.9) - [important hotfix](https://groups.google.com/forum/#!topic/orient-database/vPF85I5Blts)
+* Add health check
 
 ### 1.0.1 (2014-09-05)
 * Avoid print server config users in console help